# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurant API",
        "description": "Dokumentasi API autentikasi untuk aplikasi restoran.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/guest/midtrans/create-snap-token": {
            "post": {
                "tags": [
                    "Payments (Midtrans)"
                ],
                "summary": "Buat Snap Token untuk pembayaran",
                "description": "Generate token untuk Midtrans Snap payment gateway",
                "operationId": "f06cf76d0c2c824722b2d831fcdfdc43",
                "requestBody": {
                    "description": "Order ID yang akan dibayar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id"
                                ],
                                "properties": {
                                    "order_id": {
                                        "description": "ID order dari database",
                                        "type": "string",
                                        "example": "OR0001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Snap token berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "snap_token": {
                                            "description": "Token untuk Midtrans Snap",
                                            "type": "string"
                                        },
                                        "order_id": {
                                            "type": "string",
                                            "example": "OR0001"
                                        },
                                        "midtrans_order_id": {
                                            "type": "string",
                                            "example": "OR0001-1234567890"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal - order tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error saat membuat token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Gagal membuat token pembayaran"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/guest/midtrans/notification": {
            "post": {
                "tags": [
                    "Payments (Midtrans)"
                ],
                "summary": "Webhook notification dari Midtrans",
                "description": "Endpoint untuk menerima notifikasi pembayaran dari Midtrans. Dipanggil otomatis oleh Midtrans ketika ada perubahan status transaksi.",
                "operationId": "f1d13411d2dcedf4aafe68910f5e1b0b",
                "requestBody": {
                    "description": "Payload notifikasi dari Midtrans",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order_id": {
                                        "description": "Unique order ID dari Midtrans",
                                        "type": "string"
                                    },
                                    "transaction_id": {
                                        "description": "Transaction ID dari Midtrans",
                                        "type": "string"
                                    },
                                    "transaction_status": {
                                        "description": "Status transaksi",
                                        "type": "string",
                                        "enum": [
                                            "settlement",
                                            "capture",
                                            "pending",
                                            "deny",
                                            "cancel",
                                            "expire",
                                            "refund"
                                        ]
                                    },
                                    "payment_type": {
                                        "description": "Tipe pembayaran",
                                        "type": "string",
                                        "enum": [
                                            "qris",
                                            "bank_transfer",
                                            "gopay",
                                            "credit_card"
                                        ]
                                    },
                                    "fraud_status": {
                                        "description": "Status fraud detection",
                                        "type": "string",
                                        "enum": [
                                            "accept",
                                            "challenge",
                                            "deny"
                                        ]
                                    },
                                    "gross_amount": {
                                        "description": "Jumlah transaksi",
                                        "type": "number"
                                    },
                                    "va_numbers": {
                                        "description": "Virtual Account numbers (untuk bank_transfer)",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "bank": {
                                                    "type": "string"
                                                },
                                                "va_number": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "permata_va_number": {
                                        "description": "Permata VA number (untuk bank_transfer)",
                                        "type": "string"
                                    },
                                    "signature_key": {
                                        "description": "Hash signature untuk verifikasi",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notifikasi berhasil diproses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Gagal memproses notifikasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/guest/midtrans/check-status": {
            "get": {
                "tags": [
                    "Payments (Midtrans)"
                ],
                "summary": "Cek status pembayaran",
                "description": "Endpoint untuk frontend mengecek status pembayaran secara real-time",
                "operationId": "f31a2b88769f24ad090439e86f2ffd68",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "query",
                        "description": "Order ID yang ingin dicek statusnya",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "OR0001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status pembayaran berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "enum": [
                                                        "pending",
                                                        "paid",
                                                        "failed",
                                                        "expired"
                                                    ],
                                                    "example": "paid"
                                                },
                                                "payment_type": {
                                                    "type": "string",
                                                    "example": "qris"
                                                },
                                                "midtrans_transaction_id": {
                                                    "type": "string"
                                                },
                                                "midtrans_order_id": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment record tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Payment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error internal server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}